+++ SonicTheHedgehog/Modules/Skills.cs
@@ -12,8 +12,6 @@ namespace SonicTheHedgehog.Modules
     internal static class Skills
     {
         #region genericskills
-
-        public static GenericSkill passiveFamily;
         public static void CreateSkillFamilies(GameObject targetPrefab, bool destroyExisting = true)
         {
             if (destroyExisting)
@@ -26,8 +24,6 @@ namespace SonicTheHedgehog.Modules
 
             SkillLocator skillLocator = targetPrefab.GetComponent<SkillLocator>();
 
-
-            passiveFamily = CreateGenericSkillWithSkillFamily(targetPrefab, "Misc");
             skillLocator.primary = CreateGenericSkillWithSkillFamily(targetPrefab, "Primary");
             skillLocator.secondary = CreateGenericSkillWithSkillFamily(targetPrefab, "Secondary");
             skillLocator.utility = CreateGenericSkillWithSkillFamily(targetPrefab, "Utility");
@@ -89,10 +85,6 @@ namespace SonicTheHedgehog.Modules
         public static void AddSpecialSkills(GameObject targetPrefab, params SkillDef[] skillDefs)
         {
             AddSkillsToFamily(targetPrefab.GetComponent<SkillLocator>().special.skillFamily, skillDefs);
-        }
-        public static void AddMiscSkills(GameObject targetPrefab, params SkillDef[] skillDefs)
-        {
-            AddSkillsToFamily(passiveFamily.skillFamily, skillDefs);
         }