+++ SonicTheHedgehog/SkillStates/SonicTheHedgehog/SonicMelee.cs
@@ -12,6 +12,7 @@ using System.Linq;
 using System.Reflection;
 using UnityEngine;
 using UnityEngine.Networking;
+using static SonicTheHedgehog.Modules.SkillDefs;
 
 namespace SonicTheHedgehog.SkillStates
 {
@@ -64,7 +65,7 @@ namespace SonicTheHedgehog.SkillStates
         private bool swingSoundPlayed = false;
         private bool bufferedHomingAttack = false;
 
-        public SkillDefs.IMeleeSkill meleeSkillDef;
+        public MeleeSkillDef meleeSkillDef;
 
         public override void OnEnter()
         {
@@ -76,7 +77,7 @@ namespace SonicTheHedgehog.SkillStates
             }
             this.hasFired = false;
 
-            meleeSkillDef = base.skillLocator.primary.skillDef as SkillDefs.IMeleeSkill;
+            meleeSkillDef = base.skillLocator.primary.skillDef as MeleeSkillDef;
 
             this.impactSound = swingIndex == 4 ? Modules.Assets.meleeFinalHitSoundEvent.index : Modules.Assets.meleeHitSoundEvent.index;
             this.swingSoundString = swingIndex == 4 ? "Play_sonicthehedgehog_swing_strong" : "Play_sonicthehedgehog_swing";
@@ -347,7 +348,7 @@ namespace SonicTheHedgehog.SkillStates
             {
                 if (bufferedHomingAttack && homingTracker.CanHomingAttack())
                 {
-                    HomingAttack homingAttack = EntityStateCatalog.InstantiateState(((SkillDefs.IMeleeSkill)base.skillLocator.primary.skillDef).homingAttackState.stateType) as HomingAttack;
+                    HomingAttack homingAttack = EntityStateCatalog.InstantiateState(((MeleeSkillDef)base.skillLocator.primary.skillDef).homingAttackState.stateType) as HomingAttack;
                     homingAttack.target = homingTracker.GetTrackingTarget();
                     this.outer.SetNextState(homingAttack);
                     base.characterBody.OnSkillActivated(skillLocator.primary);